first commit to the video call app

+++ controllers/specialities.js
@@ -1,13 +1,60 @@
 const SpecialityModel = require('../models/SpecialityModel')
 const CounterModel = require('../models/CounterModel')
 const specialityValidation = require('../validations/specialities')
-
+const UserModel = require('../models/UserModel')
+const mongoose = require('mongoose')
 
 const getSpecialities = async (request, response) => {
 
     try {
 
-        const specialities = await SpecialityModel.find()
+        const specialities = await SpecialityModel.find({ type: 'MAIN' })
+
+        return response.status(200).json({
+            accepted: true,
+            specialities
+        })
+
+    } catch(error) {
+        console.error(error)
+        return response.status(500).json({
+            accepted: false,
+            message: 'internal server error',
+            error: error.message
+        })
+    }
+}
+
+const getSpeciality = async (request, response) => {
+
+    try {
+
+        const { specialityId } = request.params
+
+        const speciality = await SpecialityModel.findById(specialityId)
+
+        return response.status(200).json({
+            accepted: true,
+            speciality
+        })
+
+    } catch(error) {
+        console.error(error)
+        return response.status(500).json({
+            accepted: false,
+            message: 'internal server error',
+            error: error.message
+        })
+    }
+}
+
+const getSubSpecialitiesOfMainSpeciality = async (request, response) => {
+
+    try {
+
+        const { specialityId } = request.params
+
+        const specialities = await SpecialityModel.find({ mainSpecialityId: specialityId, type: 'SUB' })
 
         return response.status(200).json({
             accepted: true,
@@ -37,7 +84,7 @@ const addSpeciality = async (request, response) => {
             })
         }
 
-        const { name, description } = request.body
+        const { name, type, mainSpecialityId } = request.body
 
         const nameList = await SpecialityModel.find({ name })
         if(nameList.length != 0) {
@@ -48,13 +95,24 @@ const addSpeciality = async (request, response) => {
             })
         }
 
+        if(type == 'SUB') {
+            const mainSpecialityList = await SpecialityModel.findById(mainSpecialityId)
+            if(mainSpecialityList.length == 0) {
+                return response.status(400).json({
+                    accepted: false,
+                    message: 'Main speciality ID is not registered',
+                    field: 'mainSpecialityId'
+                })
+            }
+        }
+
         const counter = await CounterModel.findOneAndUpdate(
             { name: 'speciality' },
             { $inc: { value: 1 } },
             { new: true, upsert: true }
         )
 
-        const specialityData = { specialityId: counter.value, name, description }
+        const specialityData = { specialityId: counter.value, ...request.body }
         const specialityObj = new SpecialityModel(specialityData)
         const newSpeciality = await specialityObj.save()
 
@@ -80,11 +138,31 @@ const deleteSpeciality = async (request, response) => {
 
         const { specialityId } = request.params
 
+        const mainUsersList = await UserModel.find({ speciality: { $in: [mongoose.Types.ObjectId(specialityId)] } })
+
+        if(mainUsersList.length != 0) {
+            return response.status(400).json({
+                accepted: false,
+                message: 'Specialty registered with users',
+                field: 'specialityId'
+            })
+        }
+
+        const subUsersList = await UserModel.find({ subSpeciality: { $in: [mongoose.Types.ObjectId(specialityId)] } })
+
+        if(subUsersList.length != 0) {
+            return response.status(400).json({
+                accepted: false,
+                message: 'Subspecialty registered with users',
+                field: 'specialityId'
+            })
+        }
+
         const deletedSpeciality = await SpecialityModel.findByIdAndDelete(specialityId)
 
         return response.status(200).json({
             accepted: true,
-            message: 'deleted speciality successfully!',
+            message: 'deleted specialty successfully!',
             speciality: deletedSpeciality
         })
 
@@ -173,7 +251,9 @@ const updateSpeciality = async (request, response) => {
 
 module.exports = { 
     getSpecialities, 
-    addSpeciality, 
+    getSpeciality,
+    getSubSpecialitiesOfMainSpeciality,
+    addSpeciality,
     deleteSpeciality, 
     deleteSpecialities, 
     updateSpeciality
